Add layer to abstract out API implementation
This adds a framework for an interface that can be used in order to swap
out different backend API implementations. The client code shouldn't
have to worry about what API is being talked to as long as it works
through the `NetworkAdapter` interface. The only time the client needs
to be aware of what it's going to talk to is at the beginning of the app
when it needs to authenticate itself to the API. Unfortunately this step
could not be cleanly abstracted away from the client as each API has
potentially different authentication mechanisms. Once authentication is
complete we are given a `NetworkAdapter` object that we can use
throughout the rest of the app. So all customization comes at
authentication time, after that the API integration becomes opaque to
the client.

I tested out a few different ideas on creating this abstraction and this
solution seemed to me to be the simplest and easiest to implement. Some
of my other attempts included modifying the `Request` and `Response`
protocols at the `YAPI` layer to make them more customizable to support
different API integrations, but this very quickly made the interface
into the `YAPI` layer unweildy and very confusing.