Defer completion handler on simultaneous image loads
Previously, if there was an attempted image load while the same image
was already being loaded we would just return an error saying to wait
for the in progress image load to complete. This provides a pretty poor
interface as there's no notification for when the image load completes,
so the client would have to either poll the cached image or set up their
own notification to properly load the image.

This change makes it so instead of returning an error we defer the
running of the completion handler until the in progress load completes.
Once the load completes it signals all other handlers waiting for the
image and runs them.